Add additional members to queue_t structure size_t size present the size of the queue list_ele_t *tail point to the last element of the queue

+++ queue.c
@@ -79,7 +79,7 @@ bool q_remove_head(queue_t *q, char *sp, size_t bufsize)
  * Return number of elements in queue.
  * Return 0 if q is NULL or empty
  */
-int q_size(queue_t *q)
+size_t q_size(queue_t *q)
 {
     /* TODO: You need to write the code for this function */
     /* Remember: It should operate in O(1) time */
